requirements.txt
discord
web3
python-dotenv
beautifulsoup4